// src/api/models/UserRouteDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TouristServer
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoutePlaceDto } from './RoutePlaceDto';
import {
    RoutePlaceDtoFromJSON,
    RoutePlaceDtoFromJSONTyped,
    RoutePlaceDtoToJSON,
} from './RoutePlaceDto';

/**
 * 
 * @export
 * @interface UserRouteDto
 */
export interface UserRouteDto {
    /**
     * 
     * @type {string}
     * @memberof UserRouteDto
     */
    name?: string | null;
    /**
     * 
     * @type {Array<RoutePlaceDto>}
     * @memberof UserRouteDto
     */
    routePlaces?: Array<RoutePlaceDto> | null;
}

/**
 * Check if a given object implements the UserRouteDto interface.
 */
export function instanceOfUserRouteDto(value: object): value is UserRouteDto {
    return true;
}

export function UserRouteDtoFromJSON(json: any): UserRouteDto {
    return UserRouteDtoFromJSONTyped(json, false);
}

export function UserRouteDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserRouteDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'routePlaces': json['routePlaces'] == null ? undefined : ((json['routePlaces'] as Array<any>).map(RoutePlaceDtoFromJSON)),
    };
}

export function UserRouteDtoToJSON(value?: UserRouteDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'routePlaces': value['routePlaces'] == null ? undefined : ((value['routePlaces'] as Array<any>).map(RoutePlaceDtoToJSON)),
    };
}