// src/api/models/userRouteDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TouristServer
 * OpenAPI spec version: 1.0
 */
import type { RoutePlaceDto } from './routePlaceDto';

export interface UserRouteDto {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  routePlaces?: RoutePlaceDto[] | null;
}