// src/api/models/UserRouteDtoReq.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TouristServer
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoutePlaceDtoReq } from './RoutePlaceDtoReq';
import {
    RoutePlaceDtoReqFromJSON,
    RoutePlaceDtoReqFromJSONTyped,
    RoutePlaceDtoReqToJSON,
} from './RoutePlaceDtoReq';

/**
 * 
 * @export
 * @interface UserRouteDtoReq
 */
export interface UserRouteDtoReq {
    /**
     * 
     * @type {string}
     * @memberof UserRouteDtoReq
     */
    name?: string | null;
    /**
     * 
     * @type {Array<RoutePlaceDtoReq>}
     * @memberof UserRouteDtoReq
     */
    routePlaces?: Array<RoutePlaceDtoReq> | null;
}

/**
 * Check if a given object implements the UserRouteDtoReq interface.
 */
export function instanceOfUserRouteDtoReq(value: object): value is UserRouteDtoReq {
    return true;
}

export function UserRouteDtoReqFromJSON(json: any): UserRouteDtoReq {
    return UserRouteDtoReqFromJSONTyped(json, false);
}

export function UserRouteDtoReqFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserRouteDtoReq {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'routePlaces': json['routePlaces'] == null ? undefined : ((json['routePlaces'] as Array<any>).map(RoutePlaceDtoReqFromJSON)),
    };
}

export function UserRouteDtoReqToJSON(value?: UserRouteDtoReq | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'routePlaces': value['routePlaces'] == null ? undefined : ((value['routePlaces'] as Array<any>).map(RoutePlaceDtoReqToJSON)),
    };
}